Check in sketch of a v1.0.0 client
