--- arxiv/arxiv-v1.py
@@ -1,1 +1,0 @@
-# TODO: rename this file to arxiv.py
